Add styles for card component
